cmd/trace/tcp: Move parser and command definition to a common package
